+++ tasks.rb
@@ -432,7 +432,6 @@ def create_hpxmls
 
   puts "Generating #{hpxmls_files.size} HPXML files..."
 
-  hpxml_docs = {}
   hpxmls_files.each_with_index do |(hpxml_file, orig_parent), i|
     puts "[#{i + 1}/#{hpxmls_files.size}] Generating #{hpxml_file}..."
 
@@ -495,7 +494,7 @@ def create_hpxmls
         hpxml_element = XMLHelper.get_element(hpxml_doc, '/HPXML')
         building_element = XMLHelper.get_element(hpxml_element, 'Building')
         for i in 2..3
-          new_building_element = deep_copy_object(building_element)
+          new_building_element = Marshal.load(Marshal.dump(building_element)) # Deep copy
 
           # Make all IDs unique so the HPXML is valid
           new_building_element.each_node do |node|
@@ -513,7 +512,13 @@ def create_hpxmls
       end
 
       XMLHelper.write_file(hpxml_doc, hpxml_path)
-      hpxml_docs[File.basename(hpxml_file)] = deep_copy_object(hpxml_doc)
+
+      schema_path = File.join(File.dirname(__FILE__), 'HPXMLtoOpenStudio', 'resources', 'hpxml_schema', 'HPXML.xsd')
+      errors, _ = XMLValidator.validate_against_schema(hpxml_path, schema_path)
+      if errors.size > 0
+        puts "\nError: Did not successfully validate #{hpxml_file}."
+        exit!
+      end
     rescue Exception => e
       puts "\n#{e}\n#{e.backtrace.join('\n')}"
       puts "\nError: Did not successfully generate #{hpxml_file}."
@@ -539,12 +544,6 @@ def create_hpxmls
       puts "Warning: Extra HPXML file found at #{File.absolute_path(hpxml)}"
     end
   end
-
-  if hpxml_docs.size != hpxmls_files.size
-    return
-  end
-
-  return hpxml_docs
 end
 
 def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
@@ -553,7 +552,7 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
   else
     args['hpxml_path'] = "workflow/sample_files/#{hpxml_file}"
   end
-  args['apply_validation'] = false # It's faster not to validate and the CI tests will catch issues
+  args['apply_validation'] = false
 
   if ['base.xml'].include? hpxml_file
     args['simulation_control_timestep'] = 60
@@ -2840,19 +2839,19 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                      interior_adjacent_to: HPXML::LocationLivingSpace,
                      area: 550,
                      insulation_assembly_r_value: 18.7,
-                     other_space_above_or_below: HPXML::FloorOtherSpaceBelow)
+                     floor_or_ceiling: HPXML::FloorTypeFloor)
     hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
                      exterior_adjacent_to: HPXML::LocationOtherMultifamilyBufferSpace,
                      interior_adjacent_to: HPXML::LocationLivingSpace,
                      area: 200,
                      insulation_assembly_r_value: 18.7,
-                     other_space_above_or_below: HPXML::FloorOtherSpaceBelow)
+                     floor_or_ceiling: HPXML::FloorTypeFloor)
     hpxml.floors.add(id: "Floor#{hpxml.floors.size + 1}",
                      exterior_adjacent_to: HPXML::LocationOtherHeatedSpace,
                      interior_adjacent_to: HPXML::LocationLivingSpace,
                      area: 150,
                      insulation_assembly_r_value: 2.1,
-                     other_space_above_or_below: HPXML::FloorOtherSpaceBelow)
+                     floor_or_ceiling: HPXML::FloorTypeFloor)
     wall = hpxml.walls.select { |w|
              w.interior_adjacent_to == HPXML::LocationLivingSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherMultifamilyBufferSpace
@@ -2960,7 +2959,8 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                      interior_adjacent_to: HPXML::LocationLivingSpace,
                      area: 450,
                      interior_finish_type: HPXML::InteriorFinishGypsumBoard,
-                     insulation_assembly_r_value: 39.3)
+                     insulation_assembly_r_value: 39.3,
+                     floor_or_ceiling: HPXML::FloorTypeCeiling)
     hpxml.slabs[0].area = 1350
     hpxml.slabs[0].exposed_perimeter = 150
     hpxml.windows[1].area = 108
@@ -3204,7 +3204,8 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                      exterior_adjacent_to: HPXML::LocationCrawlspaceUnvented,
                      interior_adjacent_to: HPXML::LocationLivingSpace,
                      area: 675,
-                     insulation_assembly_r_value: 18.7)
+                     insulation_assembly_r_value: 18.7,
+                     floor_or_ceiling: HPXML::FloorTypeFloor)
     hpxml.slabs[0].area = 675
     hpxml.slabs[0].exposed_perimeter = 75
     hpxml.slabs.add(id: "Slab#{hpxml.slabs.size + 1}",
@@ -3345,7 +3346,8 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                      exterior_adjacent_to: HPXML::LocationGarage,
                      interior_adjacent_to: HPXML::LocationLivingSpace,
                      area: 400,
-                     insulation_assembly_r_value: 39.3)
+                     insulation_assembly_r_value: 39.3,
+                     floor_or_ceiling: HPXML::FloorTypeFloor)
     hpxml.slabs[0].area -= 400
     hpxml.slabs[0].exposed_perimeter -= 40
     hpxml.slabs.add(id: "Slab#{hpxml.slabs.size + 1}",
@@ -3782,11 +3784,13 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                                                                  duct_leakage_units: HPXML::UnitsCFM25,
                                                                  duct_leakage_value: 10,
                                                                  duct_leakage_total_or_to_outside: HPXML::DuctLeakageToOutside)
-      hpxml.hvac_distributions[-1].ducts.add(duct_type: HPXML::DuctTypeSupply,
+      hpxml.hvac_distributions[-1].ducts.add(id: "Ducts#{hpxml.hvac_distributions[-1].ducts.size + 1}",
+                                             duct_type: HPXML::DuctTypeSupply,
                                              duct_insulation_r_value: 0,
                                              duct_location: HPXML::LocationOtherMultifamilyBufferSpace,
                                              duct_surface_area: 50)
-      hpxml.hvac_distributions[-1].ducts.add(duct_type: HPXML::DuctTypeReturn,
+      hpxml.hvac_distributions[-1].ducts.add(id: "Ducts#{hpxml.hvac_distributions[-1].ducts.size + 1}",
+                                             duct_type: HPXML::DuctTypeReturn,
                                              duct_insulation_r_value: 0,
                                              duct_location: HPXML::LocationOtherMultifamilyBufferSpace,
                                              duct_surface_area: 20)
@@ -3898,7 +3902,9 @@ def apply_hpxml_modification(hpxml_file, hpxml)
          'base-enclosure-2stories-garage.xml',
          'base-hvac-ducts-area-fractions.xml'].include? hpxml_file
     hpxml.hvac_distributions[0].ducts << hpxml.hvac_distributions[0].ducts[0].dup
+    hpxml.hvac_distributions[0].ducts[-1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size}"
     hpxml.hvac_distributions[0].ducts << hpxml.hvac_distributions[0].ducts[1].dup
+    hpxml.hvac_distributions[0].ducts[-1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size}"
     hpxml.hvac_distributions[0].ducts[2].duct_location = HPXML::LocationExteriorWall
     hpxml.hvac_distributions[0].ducts[2].duct_surface_area = 37.5
     hpxml.hvac_distributions[0].ducts[3].duct_location = HPXML::LocationLivingSpace
@@ -3922,42 +3928,67 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
                                  distribution_system_type: HPXML::HVACDistributionTypeAir,
                                  air_type: HPXML::AirTypeRegularVelocity)
-    hpxml.hvac_distributions[-1].duct_leakage_measurements.add(duct_type: HPXML::DuctTypeSupply,
-                                                               duct_leakage_units: HPXML::UnitsCFM25,
-                                                               duct_leakage_value: 75,
-                                                               duct_leakage_total_or_to_outside: HPXML::DuctLeakageToOutside)
-    hpxml.hvac_distributions[-1].duct_leakage_measurements.add(duct_type: HPXML::DuctTypeReturn,
-                                                               duct_leakage_units: HPXML::UnitsCFM25,
-                                                               duct_leakage_value: 25,
-                                                               duct_leakage_total_or_to_outside: HPXML::DuctLeakageToOutside)
-    hpxml.hvac_distributions[0].ducts.add(duct_type: HPXML::DuctTypeSupply,
+    hpxml.hvac_distributions[0].duct_leakage_measurements.add(duct_type: HPXML::DuctTypeSupply,
+                                                              duct_leakage_units: HPXML::UnitsCFM25,
+                                                              duct_leakage_value: 75,
+                                                              duct_leakage_total_or_to_outside: HPXML::DuctLeakageToOutside)
+    hpxml.hvac_distributions[0].duct_leakage_measurements.add(duct_type: HPXML::DuctTypeReturn,
+                                                              duct_leakage_units: HPXML::UnitsCFM25,
+                                                              duct_leakage_value: 25,
+                                                              duct_leakage_total_or_to_outside: HPXML::DuctLeakageToOutside)
+    hpxml.hvac_distributions[0].ducts.add(id: "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}",
+                                          duct_type: HPXML::DuctTypeSupply,
                                           duct_insulation_r_value: 8,
                                           duct_location: HPXML::LocationAtticUnvented,
                                           duct_surface_area: 75)
-    hpxml.hvac_distributions[0].ducts.add(duct_type: HPXML::DuctTypeSupply,
+    hpxml.hvac_distributions[0].ducts.add(id: "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}",
+                                          duct_type: HPXML::DuctTypeSupply,
                                           duct_insulation_r_value: 8,
                                           duct_location: HPXML::LocationOutside,
                                           duct_surface_area: 75)
-    hpxml.hvac_distributions[0].ducts.add(duct_type: HPXML::DuctTypeReturn,
+    hpxml.hvac_distributions[0].ducts.add(id: "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}",
+                                          duct_type: HPXML::DuctTypeReturn,
                                           duct_insulation_r_value: 4,
                                           duct_location: HPXML::LocationAtticUnvented,
                                           duct_surface_area: 25)
-    hpxml.hvac_distributions[0].ducts.add(duct_type: HPXML::DuctTypeReturn,
+    hpxml.hvac_distributions[0].ducts.add(id: "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}",
+                                          duct_type: HPXML::DuctTypeReturn,
                                           duct_insulation_r_value: 4,
                                           duct_location: HPXML::LocationOutside,
                                           duct_surface_area: 25)
-    hpxml.hvac_distributions << hpxml.hvac_distributions[0].dup
-    hpxml.hvac_distributions[-1].id = "HVACDistribution#{hpxml.hvac_distributions.size}"
+    hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
+                                 distribution_system_type: HPXML::HVACDistributionTypeAir,
+                                 air_type: HPXML::AirTypeRegularVelocity)
+    hpxml.hvac_distributions[-1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[0].dup
+    hpxml.hvac_distributions[-1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[1].dup
+    for i in 0..3
+      hpxml.hvac_distributions[-1].ducts << hpxml.hvac_distributions[0].ducts[i].dup
+      hpxml.hvac_distributions[-1].ducts[-1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size + i + 1}"
+    end
     hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
                                  distribution_system_type: HPXML::HVACDistributionTypeHydronic,
                                  hydronic_type: HPXML::HydronicTypeBaseboard)
     hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
                                  distribution_system_type: HPXML::HVACDistributionTypeHydronic,
                                  hydronic_type: HPXML::HydronicTypeBaseboard)
-    hpxml.hvac_distributions << hpxml.hvac_distributions[0].dup
-    hpxml.hvac_distributions[-1].id = "HVACDistribution#{hpxml.hvac_distributions.size}"
-    hpxml.hvac_distributions << hpxml.hvac_distributions[0].dup
-    hpxml.hvac_distributions[-1].id = "HVACDistribution#{hpxml.hvac_distributions.size}"
+    hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
+                                 distribution_system_type: HPXML::HVACDistributionTypeAir,
+                                 air_type: HPXML::AirTypeRegularVelocity)
+    hpxml.hvac_distributions[-1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[0].dup
+    hpxml.hvac_distributions[-1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[1].dup
+    for i in 0..3
+      hpxml.hvac_distributions[-1].ducts << hpxml.hvac_distributions[0].ducts[i].dup
+      hpxml.hvac_distributions[-1].ducts[-1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size * 2 + i + 1}"
+    end
+    hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
+                                 distribution_system_type: HPXML::HVACDistributionTypeAir,
+                                 air_type: HPXML::AirTypeRegularVelocity)
+    hpxml.hvac_distributions[-1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[0].dup
+    hpxml.hvac_distributions[-1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[1].dup
+    for i in 0..3
+      hpxml.hvac_distributions[-1].ducts << hpxml.hvac_distributions[0].ducts[i].dup
+      hpxml.hvac_distributions[-1].ducts[-1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size * 3 + i + 1}"
+    end
     hpxml.heating_systems.reverse_each do |heating_system|
       heating_system.delete
     end
@@ -4080,8 +4111,15 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          primary_heating_system: true)
   elsif ['base-mechvent-multiple.xml',
          'base-bldgtype-multifamily-shared-mechvent-multiple.xml'].include? hpxml_file
-    hpxml.hvac_distributions << hpxml.hvac_distributions[0].dup
-    hpxml.hvac_distributions[1].id = "HVACDistribution#{hpxml.hvac_distributions.size}"
+    hpxml.hvac_distributions.add(id: "HVACDistribution#{hpxml.hvac_distributions.size + 1}",
+                                 distribution_system_type: HPXML::HVACDistributionTypeAir,
+                                 air_type: HPXML::AirTypeRegularVelocity)
+    hpxml.hvac_distributions[1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[0].dup
+    hpxml.hvac_distributions[1].duct_leakage_measurements << hpxml.hvac_distributions[0].duct_leakage_measurements[1].dup
+    hpxml.hvac_distributions[1].ducts << hpxml.hvac_distributions[0].ducts[0].dup
+    hpxml.hvac_distributions[1].ducts << hpxml.hvac_distributions[0].ducts[1].dup
+    hpxml.hvac_distributions[1].ducts[0].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}"
+    hpxml.hvac_distributions[1].ducts[1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size + 2}"
     hpxml.heating_systems[0].heating_capacity /= 2.0
     hpxml.heating_systems[0].fraction_heat_load_served /= 2.0
     hpxml.heating_systems[0].primary_system = false
@@ -4098,11 +4136,13 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml.cooling_systems[1].primary_system = true
   elsif ['base-bldgtype-multifamily-adjacent-to-multiple.xml'].include? hpxml_file
     hpxml.hvac_distributions[0].ducts[1].duct_location = HPXML::LocationOtherHousingUnit
-    hpxml.hvac_distributions[0].ducts.add(duct_type: HPXML::DuctTypeSupply,
+    hpxml.hvac_distributions[0].ducts.add(id: "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}",
+                                          duct_type: HPXML::DuctTypeSupply,
                                           duct_insulation_r_value: 4,
                                           duct_location: HPXML::LocationRoofDeck,
                                           duct_surface_area: 150)
-    hpxml.hvac_distributions[0].ducts.add(duct_type: HPXML::DuctTypeReturn,
+    hpxml.hvac_distributions[0].ducts.add(id: "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}",
+                                          duct_type: HPXML::DuctTypeReturn,
                                           duct_insulation_r_value: 0,
                                           duct_location: HPXML::LocationRoofDeck,
                                           duct_surface_area: 50)
@@ -4126,6 +4166,8 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml.hvac_distributions[0].duct_leakage_measurements << hpxml.hvac_distributions[1].duct_leakage_measurements[1].dup
     hpxml.hvac_distributions[0].ducts << hpxml.hvac_distributions[1].ducts[0].dup
     hpxml.hvac_distributions[0].ducts << hpxml.hvac_distributions[1].ducts[1].dup
+    hpxml.hvac_distributions[1].ducts[0].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size + 1}"
+    hpxml.hvac_distributions[1].ducts[1].id = "Ducts#{hpxml.hvac_distributions[0].ducts.size + 2}"
   end
   if ['base-hvac-ducts-area-multipliers.xml'].include? hpxml_file
     hpxml.hvac_distributions[0].ducts[0].duct_surface_area_multiplier = 0.5
@@ -4708,266 +4750,6 @@ def download_epws
   exit!
 end
 
-def get_elements_from_sample_files(hpxml_docs)
-  elements_being_used = []
-  hpxml_docs.values.each do |hpxml_doc|
-    root = XMLHelper.get_element(hpxml_doc, '/HPXML')
-    root.each_node do |node|
-      next unless node.is_a?(Oga::XML::Element)
-
-      ancestors = []
-      node.each_ancestor do |parent_node|
-        ancestors << ['h:', parent_node.name].join()
-      end
-      parent_element_xpath = ancestors.reverse
-      child_element_xpath = ['h:', node.name].join()
-      element_xpath = [parent_element_xpath, child_element_xpath].join('/')
-
-      next if element_xpath.include? 'extension'
-
-      elements_being_used << element_xpath if not elements_being_used.include? element_xpath
-    end
-  end
-
-  return elements_being_used
-end
-
-def create_schematron_hpxml_validator(hpxml_docs)
-  puts 'Generating HPXMLvalidator.xml...'
-  elements_in_sample_files = get_elements_from_sample_files(hpxml_docs)
-
-  hpxml_base_elements_xsd = File.read(File.join(File.dirname(__FILE__), 'HPXMLtoOpenStudio', 'resources', 'hpxml_schema', 'HPXMLBaseElements.xsd'))
-  hpxml_base_elements_xsd_doc = Oga.parse_xml(hpxml_base_elements_xsd)
-
-  # construct dictionary for enumerations and min/max values of HPXML data types
-  hpxml_data_types_xsd = File.read(File.join(File.dirname(__FILE__), 'HPXMLtoOpenStudio', 'resources', 'hpxml_schema', 'HPXMLDataTypes.xsd'))
-  hpxml_data_types_xsd_doc = Oga.parse_xml(hpxml_data_types_xsd)
-  hpxml_data_types_dict = {}
-  hpxml_data_types_xsd_doc.xpath('//xs:simpleType | //xs:complexType').each do |simple_type_element|
-    enums = []
-    simple_type_element.xpath('xs:restriction/xs:enumeration').each do |enum|
-      enums << enum.get('value')
-    end
-    minInclusive_element = simple_type_element.at_xpath('xs:restriction/xs:minInclusive')
-    min_inclusive = minInclusive_element.get('value') if not minInclusive_element.nil?
-    maxInclusive_element = simple_type_element.at_xpath('xs:restriction/xs:maxInclusive')
-    max_inclusive = maxInclusive_element.get('value') if not maxInclusive_element.nil?
-    minExclusive_element = simple_type_element.at_xpath('xs:restriction/xs:minExclusive')
-    min_exclusive = minExclusive_element.get('value') if not minExclusive_element.nil?
-    maxExclusive_element = simple_type_element.at_xpath('xs:restriction/xs:maxExclusive')
-    max_exclusive = maxExclusive_element.get('value') if not maxExclusive_element.nil?
-
-    simple_type_element_name = simple_type_element.get('name')
-    hpxml_data_types_dict[simple_type_element_name] = {}
-    hpxml_data_types_dict[simple_type_element_name][:enums] = enums
-    hpxml_data_types_dict[simple_type_element_name][:min_inclusive] = min_inclusive
-    hpxml_data_types_dict[simple_type_element_name][:max_inclusive] = max_inclusive
-    hpxml_data_types_dict[simple_type_element_name][:min_exclusive] = min_exclusive
-    hpxml_data_types_dict[simple_type_element_name][:max_exclusive] = max_exclusive
-  end
-
-  # construct HPXMLvalidator.xml
-  hpxml_validator = XMLHelper.create_doc('1.0', 'UTF-8')
-  root = XMLHelper.add_element(hpxml_validator, 'sch:schema')
-  XMLHelper.add_attribute(root, 'xmlns:sch', 'http://purl.oclc.org/dsdl/schematron')
-  XMLHelper.add_element(root, 'sch:title', 'HPXML Schematron Validator: HPXML.xsd', :string)
-  name_space = XMLHelper.add_element(root, 'sch:ns')
-  XMLHelper.add_attribute(name_space, 'uri', 'http://hpxmlonline.com/2019/10')
-  XMLHelper.add_attribute(name_space, 'prefix', 'h')
-  pattern = XMLHelper.add_element(root, 'sch:pattern')
-
-  # construct complexType and group elements dictionary
-  complex_type_or_group_dict = {}
-  ['//xs:complexType', '//xs:group', '//xs:element'].each do |param|
-    hpxml_base_elements_xsd_doc.xpath(param).each do |param_type|
-      next if param_type.name == 'element' && (not ['XMLTransactionHeaderInformation', 'ProjectStatus', 'SoftwareInfo'].include?(param_type.get('name')))
-      next if param_type.get('name').nil?
-
-      param_type_name = param_type.get('name')
-      complex_type_or_group_dict[param_type_name] = {}
-
-      elements = { 'child' => [], 'base' => [] }
-      param_type.each_node do |element|
-        elements['child'] << element
-        next unless element.is_a? Oga::XML::Element
-
-        next unless element.name == 'extension'
-
-        base_element_name = element.get('base').to_s
-        hpxml_base_elements_xsd_doc.xpath("#{param}[@name='#{base_element_name}']").each do |base_element|
-          base_element.each_node do |element|
-            elements['base'] << element
-          end
-        end
-      end
-
-      elements.each do |element_child_or_base, element_list|
-        element_list.each do |element|
-          next unless element.is_a? Oga::XML::Element
-          next unless (element.name == 'element' || element.name == 'group')
-          next if element.name == 'element' && (element.get('name').nil? && element.get('ref').nil?)
-          next if element.name == 'group' && element.get('ref').nil?
-
-          ancestors = []
-          element.each_ancestor do |node|
-            next if node.get('name').nil?
-            next if node.get('name') == param_type.get('name') # exclude complexType name from element xpath
-
-            ancestors << node.get('name')
-          end
-          ancestors.shift if element_child_or_base == 'base'
-
-          parent_element_names = ancestors.reverse
-          if element.name == 'element'
-            child_element_name = element.get('name')
-            child_element_name = element.get('ref') if child_element_name.nil? # Backup
-            element_type = element.get('type')
-            element_type = element.get('ref') if element_type.nil? # Backup
-          elsif element.name == 'group'
-            child_element_name = nil # exclude group name from the element's xpath
-            element_type = element.get('ref')
-          end
-          element_xpath = parent_element_names.push(child_element_name)
-          complex_type_or_group_dict[param_type_name][element_xpath] = element_type
-        end
-      end
-    end
-  end
-
-  element_xpaths = {}
-  top_level_elements_of_interest = elements_in_sample_files.map { |e| e.split('/')[1].gsub('h:', '') }.uniq
-  top_level_elements_of_interest.each do |element|
-    top_level_element = []
-    top_level_element << element
-    top_level_element_type = element
-    get_element_full_xpaths(element_xpaths, complex_type_or_group_dict, top_level_element, top_level_element_type)
-  end
-
-  # Add enumeration and min/max numeric values
-  rules = {}
-  element_xpaths.each do |element_xpath, element_type|
-    next if element_type.nil?
-
-    # Skip element xpaths not being used in sample files
-    element_xpath_with_prefix = element_xpath.compact.map { |e| "h:#{e}" }
-    context_xpath = element_xpath_with_prefix.join('/').chomp('/')
-    next unless elements_in_sample_files.any? { |item| item.include? context_xpath }
-
-    hpxml_data_type_name = [element_type, '_simple'].join() # FUTURE: This may need to be improved later since enumeration and minimum/maximum values cannot be guaranteed to always be placed within simpleType.
-    hpxml_data_type = hpxml_data_types_dict[hpxml_data_type_name]
-    hpxml_data_type = hpxml_data_types_dict[element_type] if hpxml_data_type.nil? # Backup
-    if hpxml_data_type.nil?
-      fail "Could not find data type name for '#{element_type}'."
-    end
-
-    next if hpxml_data_type[:enums].empty? && hpxml_data_type[:min_inclusive].nil? && hpxml_data_type[:max_inclusive].nil? && hpxml_data_type[:min_exclusive].nil? && hpxml_data_type[:max_exclusive].nil?
-
-    element_name = context_xpath.split('/')[-1]
-    context_xpath = context_xpath.split('/')[0..-2].join('/').chomp('/').prepend('/h:HPXML/')
-    rule = rules[context_xpath]
-    if rule.nil?
-      # Need new rule
-      rule = XMLHelper.add_element(pattern, 'sch:rule')
-      XMLHelper.add_attribute(rule, 'context', context_xpath)
-      rules[context_xpath] = rule
-    end
-
-    if not hpxml_data_type[:enums].empty?
-      assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected #{element_name.gsub('h:', '')} to be \"#{hpxml_data_type[:enums].join('" or "')}\"", :string)
-      XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-      XMLHelper.add_attribute(assertion, 'test', "#{element_name}[#{hpxml_data_type[:enums].map { |e| "text()=\"#{e}\"" }.join(' or ')}] or not(#{element_name})")
-    else
-      if hpxml_data_type[:min_inclusive]
-        assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected #{element_name.gsub('h:', '')} to be greater than or equal to #{hpxml_data_type[:min_inclusive]}", :string)
-        XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-        XMLHelper.add_attribute(assertion, 'test', "number(#{element_name}) &gt;= #{hpxml_data_type[:min_inclusive]} or not(#{element_name})")
-      end
-      if hpxml_data_type[:max_inclusive]
-        assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected #{element_name.gsub('h:', '')} to be less than or equal to #{hpxml_data_type[:max_inclusive]}", :string)
-        XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-        XMLHelper.add_attribute(assertion, 'test', "number(#{element_name}) &lt;= #{hpxml_data_type[:max_inclusive]} or not(#{element_name})")
-      end
-      if hpxml_data_type[:min_exclusive]
-        assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected #{element_name.gsub('h:', '')} to be greater than #{hpxml_data_type[:min_exclusive]}", :string)
-        XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-        XMLHelper.add_attribute(assertion, 'test', "number(#{element_name}) &gt; #{hpxml_data_type[:min_exclusive]} or not(#{element_name})")
-      end
-      if hpxml_data_type[:max_exclusive]
-        assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected #{element_name.gsub('h:', '')} to be less than #{hpxml_data_type[:max_exclusive]}", :string)
-        XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-        XMLHelper.add_attribute(assertion, 'test', "number(#{element_name}) &lt; #{hpxml_data_type[:max_exclusive]} or not(#{element_name})")
-      end
-    end
-  end
-
-  # Add ID/IDref checks
-  # FUTURE: Dynamically obtain these lists
-  id_names = ['SystemIdentifier',
-              'BuildingID']
-  idref_names = ['AttachedToRoof',
-                 'AttachedToFloor',
-                 'AttachedToSlab',
-                 'AttachedToFoundationWall',
-                 'AttachedToWall',
-                 'AttachedToRimJoist',
-                 'DistributionSystem',
-                 'AttachedToHVACDistributionSystem',
-                 'RelatedHVACSystem',
-                 'ConnectedTo']
-  elements_in_sample_files.each do |element_xpath|
-    element_name = element_xpath.split('/')[-1].gsub('h:', '')
-    context_xpath = "/#{element_xpath.split('/')[0..-2].join('/')}"
-    if id_names.include? element_name
-      rule = rules[context_xpath]
-      if rule.nil?
-        # Need new rule
-        rule = XMLHelper.add_element(pattern, 'sch:rule')
-        XMLHelper.add_attribute(rule, 'context', context_xpath)
-        rules[context_xpath] = rule
-      end
-      assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected #{element_name} with id attribute", :string)
-      XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-      XMLHelper.add_attribute(assertion, 'test', "count(h:#{element_name}[@id]) = 1")
-    elsif idref_names.include?(element_name)
-      rule = rules[context_xpath]
-      if rule.nil?
-        # Need new rule
-        rule = XMLHelper.add_element(pattern, 'sch:rule')
-        XMLHelper.add_attribute(rule, 'context', context_xpath)
-        rules[context_xpath] = rule
-      end
-      assertion = XMLHelper.add_element(rule, 'sch:assert', "Expected idref attribute for #{element_name}", :string)
-      XMLHelper.add_attribute(assertion, 'role', 'ERROR')
-      XMLHelper.add_attribute(assertion, 'test', "count(h:#{element_name}[@idref]) = count(h:#{element_name})")
-    end
-  end
-
-  XMLHelper.write_file(hpxml_validator, File.join(File.dirname(__FILE__), 'HPXMLtoOpenStudio', 'resources', 'hpxml_schematron', 'HPXMLvalidator.xml'))
-end
-
-def get_element_full_xpaths(element_xpaths, complex_type_or_group_dict, element_xpath, element_type)
-  if not complex_type_or_group_dict.keys.include? element_type
-    element_xpaths[element_xpath] = element_type
-  else
-    complex_type_or_group = deep_copy_object(complex_type_or_group_dict[element_type])
-    complex_type_or_group.each do |k, v|
-      child_element_xpath = k.unshift(element_xpath).flatten!
-      child_element_type = v
-
-      if not complex_type_or_group_dict.keys.include? child_element_type
-        element_xpaths[child_element_xpath] = child_element_type
-        next
-      end
-
-      get_element_full_xpaths(element_xpaths, complex_type_or_group_dict, child_element_xpath, child_element_type)
-    end
-  end
-end
-
-def deep_copy_object(obj)
-  return Marshal.load(Marshal.dump(obj))
-end
-
 command_list = [:update_measures, :update_hpxmls, :cache_weather, :create_release_zips, :download_weather]
 
 def display_usage(command_list)
@@ -5079,12 +4861,7 @@ if ARGV[0].to_sym == :update_hpxmls
   ENV['HOMEDRIVE'] = 'C:\\' if !ENV['HOMEDRIVE'].nil? && ENV['HOMEDRIVE'].start_with?('U:')
 
   # Create sample/test HPXMLs
-  hpxml_docs = create_hpxmls()
-
-  # Create Schematron file that reflects HPXML schema
-  if not hpxml_docs.nil?
-    create_schematron_hpxml_validator(hpxml_docs)
-  end
+  create_hpxmls()
 end
 
 if ARGV[0].to_sym == :cache_weather